requirements.txt
streamlit
pandas
plotly
kaleido
python-dotenv